+++ eva_cttv_pipeline/clinvar_xml_io/clinvar_xml_io/__init__.py
@@ -1,0 +1,1 @@
+from .clinvar_xml_io import *
