+++ eva_cttv_pipeline/clinvar_xml_io/data/test_clinvar_record.xml.gz
