--- eva_cttv_pipeline/clinvar_xml_utils/clinvar_xml_utils/__init__.py
@@ -1,1 +1,0 @@
-from .clinvar_xml_utils import *
