+++ tests/eva_cttv_pipeline/evidence_string_generation/config.py
@@ -1,6 +1,6 @@
 import os
 
-from eva_cttv_pipeline.clinvar_xml_utils import clinvar_xml_utils
+from eva_cttv_pipeline.clinvar_xml_io import clinvar_xml_io
 
 OT_SCHEMA_VERSION = "2.1.0"
 
@@ -16,4 +16,4 @@ def get_expected_evidence_string(filename):
 def get_test_clinvar_record(filename='test_clinvar_record.xml.gz'):
     """The default test file contains an extract of ClinVar XML for the record RCV000002127."""
     test_clinvar_record_file = os.path.join(test_dir, 'resources', filename)
-    return [r for r in clinvar_xml_utils.ClinVarDataset(test_clinvar_record_file)][0]
+    return [r for r in clinvar_xml_io.ClinVarDataset(test_clinvar_record_file)][0]
